Move all game state update code to the event space
Events determine how game state including the box score changes.
Lessens the somewhat awkward interactions where some of it
happened in simulation and some of it happened in the event.

Now it happens in Event and right after in the specific events
which seems closer together.

Also, they now only just update the game state as a whole and
not separately game state and statistcs as the latter is part of
the former.
We used to have some events that implemented update_statistics
as no op now they visibly just don't touch it.

Oh and the trigger for all of this was steal being able to update
statistics both for the stealer and the person it was stolen from.